Сделал в главной странице меню
Сделал в главной странице меню и добавил в путях auth/* для аунтификации
