+++ src/pages/Layout.tsx
@@ -2,7 +2,6 @@ import Head from 'next/head';
 import { Inter } from 'next/font/google';
 import { FC, ReactNode } from 'react';
 
-
 interface ILayout {
   children: ReactNode;
   title: string;
