+++ src/pages/index.tsx
@@ -1,15 +1,20 @@
-import Head from 'next/head';
-import Image from 'next/image';
-import { Inter } from 'next/font/google';
-import styles from '@/styles/Home.module.css';
 import { Layout } from './Layout';
-
-// const inter = Inter({ subsets: ['latin'] });
+import Link from 'next/link';
+import { routerConstants } from '@/Constants/RouterConstants';
 
 export default function Home() {
   return (
     <>
-      <Layout title="Project">Главная</Layout>
+      <Layout title="Project">
+        <div className="mainContainer">
+          <nav className="nav">
+            <Link href={routerConstants.LOGIN}>Вход</Link>
+            <Link href={routerConstants.REGISTRATION}>Регистрация</Link>
+            <Link href={routerConstants.FORGOT_PASSWORD}>Забыли пароль?</Link>
+            <Link href={routerConstants.CONFIRM_MOBILE}>CONFIRM_MOBILE</Link>
+          </nav>
+        </div>
+      </Layout>
     </>
   );
 }
