+++ src/styles/globals.css
@@ -1,3 +1,29 @@
 :root {
   $reds: #ff8181;
 }
+
+.mainContainer {
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.nav {
+  border: 1px solid #333;
+  padding: 5px 0;
+}
+
+.nav a {
+  text-decoration: none;
+  padding: 10px 20px;
+  transition: all 0.2s ease-in-out;
+  font-size: 20px;
+  color: #333;
+}
+
+.nav a:hover {
+  background-color: #333;
+  color: #fff;
+}
