+++ src/styles/main.scss
@@ -2,81 +2,81 @@
 @import 'reset';
 
 .background-auth-wrapper {
-	background: rgb(255, 129, 129);
-	background: linear-gradient(
-		122deg,
-		rgba(255, 129, 129, 1) 0%,
-		rgba(255, 129, 129, 1) 56%,
-		rgba(255, 255, 255, 1) 100%
-	);
-	width: 100%;
-	height: 100vh;
-	position: absolute;
-	top: 0;
-	bottom: 0;
-	z-index: -1;
+  background: rgb(255, 129, 129);
+  background: linear-gradient(
+    122deg,
+    rgba(255, 129, 129, 1) 0%,
+    rgba(255, 129, 129, 1) 56%,
+    rgba(255, 255, 255, 1) 100%
+  );
+  width: 100%;
+  height: 100vh;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  z-index: -1;
 }
 
 .auth-container {
-	width: 100%;
-	height: 500px;
-	max-width: 374px;
-	background: $white;
-	border: 0.7px solid #454545;
-	border-radius: 10px;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-	align-items: center;
-	padding-left: 10px;
-	padding-right: 10px;
-	padding-bottom: 40px;
+  width: 100%;
+  height: 600px;
+  max-width: 374px;
+  background: $white;
+  border: 0.7px solid #454545;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding-left: 10px;
+  padding-right: 10px;
+  padding-bottom: 40px;
 
-	h1 {
-		margin-top: 36px;
-		margin-bottom: 44px;
-		font-weight: 500;
-		font-size: 32px;
-		line-height: 39px;
-		color: $black-light;
-	}
+  h1 {
+    margin-top: 36px;
+    margin-bottom: 44px;
+    font-weight: 500;
+    font-size: 32px;
+    line-height: 39px;
+    color: $black-light;
+  }
 
-	span {
-		font-weight: 500;
-		font-size: 20px;
-		line-height: 24px;
-		text-align: center;
-		color: $black-light;
-	}
+  span {
+    font-weight: 500;
+    font-size: 20px;
+    line-height: 24px;
+    text-align: center;
+    color: $black-light;
+  }
 
-	a {
-		font-size: 20px;
-		font-weight: 500;
-		line-height: 24px;
-		text-decoration: none;
-		color: $black-light;
+  a {
+    font-size: 20px;
+    font-weight: 500;
+    line-height: 24px;
+    text-decoration: none;
+    color: $black-light;
 
-		&:hover {
-			color: $red;
-		}
-	}
+    &:hover {
+      color: $red;
+    }
+  }
 }
 
 @media (max-width: 1100px) {
-	.auth-container {
-		margin: 10px;
-	}
+  .auth-container {
+    margin: 10px;
+  }
 }
 
 @media (max-width: 1000px) {
-	.background-auth-wrapper {
-		background: rgb(232, 80, 80);
-		background: radial-gradient(
-			circle,
-			rgba(232, 80, 80, 0.3113620448179272) 3%,
-			rgba(255, 129, 129, 0.6474964985994398) 51%,
-			rgba(255, 255, 255, 1) 100%
-		);
-		filter: blur(250px);
-	}
+  .background-auth-wrapper {
+    background: rgb(232, 80, 80);
+    background: radial-gradient(
+      circle,
+      rgba(232, 80, 80, 0.3113620448179272) 3%,
+      rgba(255, 129, 129, 0.6474964985994398) 51%,
+      rgba(255, 255, 255, 1) 100%
+    );
+    filter: blur(250px);
+  }
 }
